+++ webui/src/shared/index.ts
@@ -12,3 +12,4 @@ export * from "./config/presets";
 export * from "./config/limits";
 export * from "./config/navigation";
 export * from "./lib/color";
+export * from "./lib/log";